requirements.txt
sentence-transformers==2.7.0
faiss-cpu==1.8.0
groq==0.9.0
streamlit==1.35.0
pandas==2.2.2
numpy==1.26.4
tqdm==4.66.4
